Add function to average n readings
Examples currently use `refresh()` to fill the data into a `hackAIRData`
struct, extract the data into separate variables to calculate the mean
value and then put the data back into the struct to use the submit function.
This is kinda clunky and this change aims to provide a better API for something
commonly used.