on error, retry getting valid data from sensor
I tested with my SDS011 sensor. In error case, the hackAIR system interprets the data as a transmitted 0µg measurement for both PM10 and PM25. This leads to an unrealistic data base!
Fix: For these cases and as a general improvement of the example, lets examine the error code!
Note: presumably, the Wemos D1 example should be modified also!